add new feature

+++ frontend/src/components/favorite.jsx
@@ -1,37 +1,29 @@
-import React, { useContext, useEffect, useRef } from 'react';
+import React, { useContext, useEffect } from 'react';
 import { AppContext } from '../appContext';
-import { Button } from '@mui/material';
-import { toast } from 'react-toastify';
+import { useSnackbar } from 'notistack';
 
 const Favorite = () => {
   const { favorites, darkMode, setFavorites, user } = useContext(AppContext);
-  const toastIdRef = useRef(null);
+  const { enqueueSnackbar } = useSnackbar();
 
   const fetchFavorites = async () => {
     if (!user?.uid) {
-      toastIdRef.current = toast.error("Please log in to your account");
+      enqueueSnackbar('Please create your account', { variant: 'warning' });
       return;
     }
-    try {
-      const response = await fetch(`http://localhost:3000/favorites?uuid=${user.uid}`, {
-        method: 'GET',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-      });
 
-      if (!response.ok) {
-        throw new Error(`Failed to fetch favorites: ${response.status} ${response.statusText}`);
-      }
+    const response = await fetch(`http://localhost:3000/favorites?uuid=${user.uid}`, {
+      method: 'GET',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+    });
 
+    if (response.ok) {
       const result = await response.json();
       if (Array.isArray(result.favorites)) {
         setFavorites(result.favorites.filter(fav => fav && fav.location));
-      } else {
-        toastIdRef.current = toast.error('Invalid data structure received');
       }
-    } catch (error) {
-      toastIdRef.current = toast.error(`Failed to fetch favorites: ${error.message}`);
     }
   };
 
@@ -40,56 +32,43 @@ const Favorite = () => {
   }, [user, setFavorites]);
 
   const handleDeleteFavorite = async (favoriteLocation) => {
-    try {
-      const response = await fetch('http://localhost:3000/favorites', {
-        method: 'DELETE',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({
-          uuid: user.uid,
-          favoriteLocation,
-        }),
-      });
+    if (!user) {
+      enqueueSnackbar('Please log in to delete favorites', { variant: 'warning' });
+      return;
+    }
 
-      if (!response.ok) {
-        if (response.status === 404) {
-          toastIdRef.current = toast.error("Favorite not found");
-        } else {
-          throw new Error(`Failed to delete favorite: ${response.status} ${response.statusText}`);
-        }
-      }
+    const response = await fetch('http://localhost:3000/favorites', {
+      method: 'DELETE',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify({
+        uuid: user.uid,
+        favoriteLocation,
+      }),
+    });
 
-      toastIdRef.current = toast.success('Favorite deleted successfully');
-
-      // Refresh the favorites list
-      fetchFavorites();
-    } catch (error) {
-      toastIdRef.current = toast.error(`Failed to delete favorite: ${error.message}`);
+    if (response.ok) {
+      enqueueSnackbar('Favorite deleted successfully', { variant: 'success' });
+      fetchFavorites(); // Refresh the favorites list
+    } else {
+      enqueueSnackbar('Failed to delete favorite', { variant: 'error' });
     }
   };
 
-  useEffect(() => {
-    return () => {
-      if (toastIdRef.current) {
-        toast.dismiss(toastIdRef.current);
-      }
-    };
-  }, [toastIdRef]);
-
   return (
-    <div className={`p-5  min-h-screen ${darkMode ? 'bg-gray-800 text-white' : 'bg-white text-gray-900'}`}>
+    <div className={`p-5 min-h-screen ${darkMode ? 'bg-gray-800 text-white' : 'bg-white text-gray-900'}`}>
       <h2 className="text-2xl font-bold text-center mb-4">Your Favorites</h2>
       {favorites.length === 0 ? (
         <p>No favorites added yet.</p>
       ) : (
-        <ul className="list-disc pl-5 space-y-4">
+        <ul className="flex flex-wrap items-center justify-center gap-4">
           {favorites.map((favorite, index) => {
             if (favorite && favorite.location) {
               return (
                 <li
                   key={index}
-                  className="mb-2 flex items-center justify-between p-4 bg-gray-200 rounded-lg shadow-md"
+                  className="mb-2 w-full md:w-1/4 flex items-center justify-between p-4 bg-gray-200 rounded-lg shadow-md"
                 >
                   <div className="flex flex-col">
                     <h1 className="text-lg font-semibold text-gray-800">
@@ -101,14 +80,12 @@ const Favorite = () => {
                       {favorite.description}
                     </h2>
                   </div>
-                  <Button
-                    variant="contained"
-                    color="secondary"
+                  <button
                     onClick={() => handleDeleteFavorite(favorite.location)}
-                    className="ml-4"
+                    className="text-white bg-gradient-to-r from-red-500 via-red-600 to-red-700 hover:bg-gradient-to-br focus:ring-4 focus:outline-none focus:ring-blue-300 dark:focus:ring-blue-800 shadow-lg shadow-blue-500/50 dark:shadow-lg dark:shadow-blue-800/80 font-medium rounded-lg text-sm px-5 py-2.5 text-center me-2 mb-2"
                   >
                     Delete
-                  </Button>
+                  </button>
                 </li>
               );
             } else {
